requirements.txt
Django>=4.2.0
django-auditlog>=3.0.0
gunicorn>=20.1.0
whitenoise>=6.0.0
psycopg2-binary>=2.9.0
dj-database-url>=2.0.0